requirements.txt
PyYAML
numpy
pandas
pprintpp
requests
tqdm
scipy
matplotlib
